合并main

+++ dal/db/video.go
@@ -2,7 +2,7 @@
  * @Author: zy 953725892@qq.com
  * @Date: 2023-01-19 11:27:35
  * @LastEditors: zy 953725892@qq.com
- * @LastEditTime: 2023-01-19 14:22:42
+ * @LastEditTime: 2023-02-01 22:25:27
  * @FilePath: /ByteCamp/dal/db/video.go
  * @Description: 视频实体类及相关crud
  *
@@ -21,13 +21,12 @@ import (
 
 type Video struct {
 	gorm.Model
-	AuthorId   int
-	PlayUrl    string
-	CoverUrl   string
-	UploadTime time.Time
-	Title      string
-	Version    optimisticlock.Version
-	// TODO: 暂时增加了视频的点赞和评论数目字段
-	FavoriteCount int
-	CommentCount  int
+	AuthorId      int
+	PlayUrl       string
+	CoverUrl      string
+	UploadTime    time.Time
+	Title         string
+	Version       optimisticlock.Version
+	FavoriteCount int64
+	CommentCount  int64
 }
